refactors local secret, adds env var imp
