+++ secretstores/local/file/filestore.go
@@ -3,7 +3,7 @@
 // Licensed under the MIT License.
 // ------------------------------------------------------------
 
-package localsecretstore
+package file
 
 import (
 	"encoding/json"
