+++ secretstores/local/file/filestore_test.go
@@ -2,7 +2,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 // ------------------------------------------------------------
-package localsecretstore
+package file
 
 import (
 	"fmt"
